requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
faiss-cpu==1.7.3  
numpy==1.24.2
openai==0.27.0
